DAH-2497 styling updates
