+++ app/javascript/modules/listingDetailsAside/ListingDetailsApply.tsx
@@ -16,11 +16,9 @@ import {
   isHabitatListing,
   isOpen,
   isRental,
-  isSale,
   paperApplicationURLs,
 } from "../../util/listingUtil"
-import { getSfGovUrl, localizedFormat, renderInlineMarkup } from "../../util/languageUtil"
-import { getHousingCounselorsPath } from "../../util/routeUtil"
+import { getSfGovUrl, renderInlineMarkup } from "../../util/languageUtil"
 import "./ListingDetailsApply.scss"
 
 export interface ListingDetailsApplyProps {
@@ -189,40 +187,10 @@ export const ListingDetailsApply = ({ listing }: ListingDetailsApplyProps) => {
     </>
   )
 
-  const needHelpBlock = (
-    <SidebarBlock title={t("listings.apply.needHelp")}>
-      {isListingRental && (
-        <div className={"mb-4"}>{t("listings.apply.visitAHousingCounselor")}</div>
-      )}
-      <LinkButton
-        transition={true}
-        newTab={true}
-        href={
-          !isListingRental
-            ? "https://www.homeownershipsf.org/buyerapplications/"
-            : getHousingCounselorsPath()
-        }
-        className={"w-full"}
-      >
-        {isListingRental
-          ? t("housingCounselor.findAHousingCounselor")
-          : t("listings.apply.visitHomeownershipSf")}
-      </LinkButton>
-    </SidebarBlock>
-  )
-
-  const expectedMoveInDateBlock = (
-    <SidebarBlock title={t("listings.expectedMoveinDate")}>
-      {localizedFormat(listing.Expected_Move_in_Date, "MMMM YYYY")}
-    </SidebarBlock>
-  )
-
   return (
     <div className="md:px-0 px-2">
       {howToApplyBlock}
       {!isFcfsBmrSales && acceptingPaperApps && submitPaperApplicationBlocks}
-      {needHelpBlock}
-      {isSale(listing) && listing.Expected_Move_in_Date && expectedMoveInDateBlock}
     </div>
   )
 }
